Update Xor.hpp
Among the adjustments I made were the following:

Wrapped the implementation of the decrypt function in a std::function object and placed it in the XorString class. This enables the decrypt function's implementation to be altered at runtime.
Because it does not affect the XorString object, I made the decrypt method a const member function.

Changed the decryptionKey argument of the decode function to reflect that it is a different key from the one used to encrypt the string.
Return enc.decrypt(key); was changed to return enc.decrypt(key) const; in the ENCRYPTED STRING macro. This is because the decrypt method is now a const member function and does not require the const keyword to be invoked.